Reimplement accessor reference getters as default interface methods
